app/requirements.txt
python-dotenv
streamlit==1.28.1
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1
langchain
openai==0.28.1
tiktoken
markdownify